Fixed type

+++ PoGo.NecroBot.Logic/Event/PokemonEvolveEvent.cs
@@ -11,7 +11,7 @@ namespace PoGo.NecroBot.Logic.Event
     {
         public int Exp;
         public PokemonId Id;
-        public uint UniqueId;
+        public ulong UniqueId;
         public EvolvePokemonResponse.Types.Result Result;
     }
 }
